refactor(api): move buildPaths function to shared library
- Move buildPaths function from app/api/sign/route.ts to lib/build-paths.ts
- Remove unused imports and dependencies from route file
- Add proper typing and documentation for the shared function
- Update tests to import from new location

fix(preview): remove DOM-dependent path calculation

- Remove measureRef and DOM-based path length calculations
- Use consistent server-side compatible implementation
- Fix dependency array in useEffect hook

feat(i18n): add landing page translations

- Add translation keys for landing page components
- Support both English and Chinese locales
- Implement useI18n hook in landing and embed components
